requirements.txt
locationsharinglib==5.0.3
PyYAML==6.0.2
sqlmodel==0.0.24
